requirements.txt
asgiref==2.3.2
astroid==2.2.3
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
dataclasses==0.6
Django==2.1.7
django-cors-headers==2.4.0
djangorestframework==3.9.2
djongo==1.2.32
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
isort==4.3.10
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
PyHamcrest==1.9.0
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
pymongo==3.7.2
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rope==0.12.0
six==1.12.0
slackclient==1.3.1
sqlparse==0.2.4
tweepy==3.7.0
txaio==18.8.1
typed-ast==1.3.1
urllib3==1.24.1
virtualenv==16.4.3
websocket-client==0.54.0
wrapt==1.11.1
zope.interface==4.6.0
